about and terms in sessionStorage, and some cleanup
