+++ src/App.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import ScrollToTop from "./lib/ScrollToTop";
+import { fetchSiteDetails } from "./lib/fetchTools";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import HomePage from "./pages/HomePage";
@@ -23,51 +24,8 @@ class App extends Component {
     };
   }
 
-  async fetchSiteDetails(siteName) {
-    let response = null;
-    let data = null;
-    try {
-      data = JSON.parse(
-        sessionStorage.getItem(`${siteName.toLowerCase()}_config`)
-      );
-    } catch (error) {
-      console.log("Site details not in storage");
-    }
-    if (data === null) {
-      console.log("Fetching site details");
-      try {
-        response = await fetch(
-          `${process.env.REACT_APP_CONFIG_PATH}/${siteName.toLowerCase()}.json`
-        );
-        data = await response.json();
-      } catch (error) {
-        console.error(`Error fetching config file`);
-        console.error(error);
-      }
-      if (data === null) {
-        try {
-          response = await fetch(
-            `${process.env.REACT_APP_CONFIG_PATH}/default.json`
-          );
-          data = await response.json();
-        } catch (error) {
-          console.error("Error fetching default.json");
-          console.error(error);
-        }
-      } else {
-        sessionStorage.setItem(
-          `${siteName.toLowerCase()}_config`,
-          JSON.stringify(data)
-        );
-      }
-    }
-    this.setState({
-      siteDetails: data
-    });
-  }
-
   componentDidMount() {
-    this.fetchSiteDetails(process.env.REACT_APP_REP_TYPE);
+    fetchSiteDetails(this, process.env.REACT_APP_REP_TYPE);
   }
 
   render() {
